+++ Sources/HTTPLayer/Request/HttpUploadRequest.swift
@@ -25,8 +25,8 @@ class HttpUploadRequest: HttpRequest {
 
      - Returns: An initialized a HttpUploadRequest object.
      */
-    init(url: URL, method: HttpMethod, resourceUrl: URL, headers: [HttpHeader]? = nil, useProgress: Bool = true) {
+    init(url: URL, method: HttpMethod, resourceUrl: URL, headers: [HttpHeader]? = nil) {
         self.resourceUrl = resourceUrl
-        super.init(url: url, method: method, headers: headers, useProgress: useProgress)
+        super.init(url: url, method: method, headers: headers)
     }
 }