+++ Tests/Unit/HTTPLayer/Request/HttpDownloadRequestTests.swift
@@ -21,20 +21,19 @@ class HttpDownloadRequestTests: XCTestCase {
     func testFullConstructor() {
         let url = rootURL.appendingPathComponent("posts/1")
         let destination = TestData.Url.fileDestination
-        let request = HttpDownloadRequest(url: url, destinationUrl: destination, useProgress: true)
+        let request = HttpDownloadRequest(url: url, destinationUrl: destination)
 
         XCTAssertEqual(request.url, url)
         XCTAssertEqual(request.method, .get)
         XCTAssertEqual(request.destinationUrl, destination)
-        XCTAssertNotNil(request.progress)
     }
 
     func testHashValue() {
         let url = rootURL.appendingPathComponent("posts/1")
         let destination1 = TestData.Url.fileDestination
         let destination2 = TestData.Url.anotherFileDestination
-        let request1 = HttpDownloadRequest(url: url, destinationUrl: destination1, useProgress: true)
-        let request2 = HttpDownloadRequest(url: url, destinationUrl: destination2, useProgress: true)
+        let request1 = HttpDownloadRequest(url: url, destinationUrl: destination1)
+        let request2 = HttpDownloadRequest(url: url, destinationUrl: destination2)
 
         XCTAssertTrue(request1.hashValue == request1.hashValue)
         XCTAssertFalse(request1.hashValue == request2.hashValue)